+++ pytest_play/providers.py
@@ -3,61 +3,6 @@ from time import sleep
 from selenium.webdriver.common.keys import Keys
 
 
-def register_steps(data):
-    """
-        You can register a group of actions as a pytest play provider and
-        **include** them in other scenario for improved reusability.
-
-        For example let's pretend we want to reuse the login steps coming from
-        a ``login.json`` file::
-
-            import pytest
-
-
-            @pytest.fixture(autouse=True)
-            def login_procedure(play_json, data_getter):
-                data = data_getter('/my/path/etc', 'login.json')
-                play_json.register_command_provider(
-                    register_steps(data), 'login.json')
-
-            def test_like(play_json, data_getter):
-                data = data_getter('/my/path/etc', 'like.json')
-                play_json.execute(data)
-
-
-        where ``like.json`` contains the steps coming from the included
-        ``login.json`` file plus additional actions::
-
-            {
-                "steps": [
-                        {
-                                "provider": "login.json"
-                                "type": "include"
-                        },
-                        {
-                                "type": "clickElement",
-                                "locator": {
-                                        "type": "css",
-                                        "value": ".like"
-                                }
-                        }
-                ]
-            }
-
-
-        **NOTE WELL**: it's up to you avoid recursion issues.
-    """
-    class PlayEngineWrapperProvider(object):
-        """ PlayEngine wrapper """
-
-        def __init__(self, engine):
-            self.engine = engine
-
-        def command_include(self, command):
-            self.engine.execute(data)
-    return PlayEngineWrapperProvider
-
-
 class SplinterCommandProvider(object):
     """ JSON executor """