+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import os
 import pytest
 import mock
 
@@ -5,10 +6,28 @@ import mock
 pytest_plugins = 'pytester'
 
 
+@pytest.fixture
+def data_base_path():
+    """ selenium/splinter base path, where json files live """
+    here = os.path.abspath(os.path.dirname(__file__))
+    return os.path.join(here, 'data')
+
+
 @pytest.fixture(scope='session')
 def variables(skin):
-    return {'skins': {skin: {'base_url': 'http://',
-                             'credentials': {}}}}
+    return {
+        'skins': {
+            skin: {
+                'base_url': 'http://',
+                'credentials': {
+                    'Administrator': {
+                        'username': 'admin',
+                        'password': 'pwd'
+                    }
+                }
+            }
+        }
+    }
 
 
 @pytest.fixture
@@ -20,3 +39,10 @@ def dummy_executor(parametrizer_class, navigation):
 @pytest.fixture
 def page_instance():
     return mock.MagicMock()
+
+
+@pytest.fixture(autouse=True)
+def included_scenario(play_json, data_getter, data_base_path):
+    data = data_getter(data_base_path, 'included-scenario.json')
+    play_json.register_steps(
+        data, 'included-scenario.json')