+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changelog
 0.0.2 (unreleased)
 ==================
 
-- Nothing changed yet.
+- implement reusable steps (include scenario)
 
+- minor documentation changes
 
 0.0.1 (2017-12-20)
 ==================